backend/requirements.txt
openai
pymupdf
dacite
python-dotenv
google-generativeai
flask
flask-cors